+++ src/test/java/org/cicirello/math/rand/RandomIndexerSampleTests.java
@@ -541,7 +541,7 @@ public class RandomIndexerSampleTests {
 
     for (int n = 3; n <= 6; n++) {
       for (int i = 0; i < 10; i++) {
-        int[] result = RandomIndexer.nextIntTriple(n, null, true);
+        int[] result = RandomIndexer.nextSortedIntTriple(n, (int[]) null);
         assertEquals(3, result.length);
         assertNotEquals(result[0], result[1]);
         assertNotEquals(result[2], result[1]);
@@ -554,9 +554,9 @@ public class RandomIndexerSampleTests {
     }
 
     int[] expected = new int[3];
-    int[] actual = RandomIndexer.nextIntTriple(5, expected, false);
+    int[] actual = RandomIndexer.nextSortedIntTriple(5, expected);
     assertTrue(expected == actual);
-    actual = RandomIndexer.nextIntTriple(5, new int[2], false);
+    actual = RandomIndexer.nextSortedIntTriple(5, new int[2]);
     assertEquals(3, actual.length);
 
     if (DISABLE_CHI_SQUARE_TESTS) return;
@@ -566,7 +566,7 @@ public class RandomIndexerSampleTests {
         int[][][] buckets = new int[n][n][n];
         int numBuckets = n * (n - 1) * (n - 2) / 6;
         for (int i = 0; i < REPS_PER_BUCKET * numBuckets; i++) {
-          int[] result = RandomIndexer.nextIntTriple(n, null, true);
+          int[] result = RandomIndexer.nextSortedIntTriple(n, (int[]) null);
           buckets[result[0]][result[1]][result[2]]++;
         }
         double chi = chiSquare(buckets, numBuckets);
@@ -626,7 +626,7 @@ public class RandomIndexerSampleTests {
 
     for (int n = 3; n <= 6; n++) {
       for (int i = 0; i < 10; i++) {
-        int[] result = RandomIndexer.nextIntTriple(n, null, true, gen);
+        int[] result = RandomIndexer.nextSortedIntTriple(n, null, gen);
         assertEquals(3, result.length);
         assertNotEquals(result[0], result[1]);
         assertNotEquals(result[2], result[1]);
@@ -643,7 +643,7 @@ public class RandomIndexerSampleTests {
         int[][][] buckets = new int[n][n][n];
         int numBuckets = n * (n - 1) * (n - 2) / 6;
         for (int i = 0; i < REPS_PER_BUCKET * numBuckets; i++) {
-          int[] result = RandomIndexer.nextIntTriple(n, null, true, gen);
+          int[] result = RandomIndexer.nextSortedIntTriple(n, null, gen);
           buckets[result[0]][result[1]][result[2]]++;
         }
         double chi = chiSquare(buckets, numBuckets);
@@ -654,9 +654,9 @@ public class RandomIndexerSampleTests {
     }
 
     int[] expected = new int[3];
-    int[] actual = RandomIndexer.nextIntTriple(5, expected, false, gen);
+    int[] actual = RandomIndexer.nextSortedIntTriple(5, expected, gen);
     assertTrue(expected == actual);
-    actual = RandomIndexer.nextIntTriple(5, new int[2], false, gen);
+    actual = RandomIndexer.nextSortedIntTriple(5, new int[2], gen);
     assertEquals(3, actual.length);
   }
 
@@ -779,7 +779,7 @@ public class RandomIndexerSampleTests {
 
     for (int n = 3; n <= 6; n++) {
       for (int i = 0; i < 10; i++) {
-        IndexTriple result = RandomIndexer.nextIntTriple(n, true);
+        IndexTriple result = RandomIndexer.nextSortedIntTriple(n);
         assertNotEquals(result.i(), result.j());
         assertNotEquals(result.k(), result.j());
         assertNotEquals(result.i(), result.k());
@@ -797,7 +797,7 @@ public class RandomIndexerSampleTests {
         int[][][] buckets = new int[n][n][n];
         int numBuckets = n * (n - 1) * (n - 2) / 6;
         for (int i = 0; i < REPS_PER_BUCKET * numBuckets; i++) {
-          IndexTriple result = RandomIndexer.nextIntTriple(n, true);
+          IndexTriple result = RandomIndexer.nextSortedIntTriple(n);
           buckets[result.i()][result.j()][result.k()]++;
         }
         double chi = chiSquare(buckets, numBuckets);
@@ -846,43 +846,6 @@ public class RandomIndexerSampleTests {
   }
 
   @Test
-  public void testTriple_IndexTriple_SortIsFalse_SplittableRandom() {
-    SplittableRandom gen = new SplittableRandom(42);
-    final int REPS_PER_BUCKET = 100;
-    final int TRIALS = 100;
-
-    for (int n = 3; n <= 6; n++) {
-      for (int i = 0; i < 10; i++) {
-        IndexTriple result = RandomIndexer.nextIntTriple(n, false, gen);
-        assertNotEquals(result.i(), result.j());
-        assertNotEquals(result.k(), result.j());
-        assertNotEquals(result.i(), result.k());
-        assertTrue(result.i() >= 0);
-        assertTrue(result.i() < n);
-        assertTrue(result.j() >= 0);
-        assertTrue(result.j() < n);
-        assertTrue(result.k() >= 0);
-        assertTrue(result.k() < n);
-      }
-    }
-    for (int n = 3; n <= 6; n++) {
-      int countH = 0;
-      for (int trial = 0; trial < TRIALS; trial++) {
-        int[][][] buckets = new int[n][n][n];
-        int numBuckets = n * (n - 1) * (n - 2);
-        for (int i = 0; i < REPS_PER_BUCKET * numBuckets; i++) {
-          IndexTriple result = RandomIndexer.nextIntTriple(n, false, gen);
-          buckets[result.i()][result.j()][result.k()]++;
-        }
-        double chi = chiSquareAll(buckets, numBuckets);
-        if (chi > limit95[numBuckets - 1]) countH++;
-      }
-      assertTrue(
-          countH <= TRIALS * 0.1, "chi square too high too often, countHigh=" + countH + " n=" + n);
-    }
-  }
-
-  @Test
   public void testTripleSorted_IndexTriple_SplittableRandom() {
     SplittableRandom gen = new SplittableRandom(42);
     final int REPS_PER_BUCKET = 100;
@@ -890,7 +853,7 @@ public class RandomIndexerSampleTests {
 
     for (int n = 3; n <= 6; n++) {
       for (int i = 0; i < 10; i++) {
-        IndexTriple result = RandomIndexer.nextIntTriple(n, true, gen);
+        IndexTriple result = RandomIndexer.nextSortedIntTriple(n, gen);
         assertNotEquals(result.i(), result.j());
         assertNotEquals(result.k(), result.j());
         assertNotEquals(result.i(), result.k());
@@ -906,7 +869,7 @@ public class RandomIndexerSampleTests {
         int[][][] buckets = new int[n][n][n];
         int numBuckets = n * (n - 1) * (n - 2) / 6;
         for (int i = 0; i < REPS_PER_BUCKET * numBuckets; i++) {
-          IndexTriple result = RandomIndexer.nextIntTriple(n, true, gen);
+          IndexTriple result = RandomIndexer.nextSortedIntTriple(n, gen);
           buckets[result.i()][result.j()][result.k()]++;
         }
         double chi = chiSquare(buckets, numBuckets);
@@ -1248,7 +1211,7 @@ public class RandomIndexerSampleTests {
         assertTrue(result[0] >= 0);
         assertTrue(result[2] < n);
         assertTrue(result[2] - result[0] <= w);
-        result = RandomIndexer.nextWindowedIntTriple(n, w, (int[]) null, true);
+        result = RandomIndexer.nextSortedWindowedIntTriple(n, w, (int[]) null);
         assertEquals(3, result.length);
         assertNotEquals(result[0], result[1]);
         assertNotEquals(result[0], result[2]);
@@ -1269,9 +1232,9 @@ public class RandomIndexerSampleTests {
     actual = RandomIndexer.nextWindowedIntTriple(5, 3, new int[2]);
     assertEquals(3, actual.length);
 
-    actual = RandomIndexer.nextWindowedIntTriple(5, 3, expected, false);
+    actual = RandomIndexer.nextSortedWindowedIntTriple(5, 3, expected);
     assertTrue(expected == actual);
-    actual = RandomIndexer.nextWindowedIntTriple(5, 3, new int[2], false);
+    actual = RandomIndexer.nextSortedWindowedIntTriple(5, 3, new int[2]);
     assertEquals(3, actual.length);
 
     if (DISABLE_CHI_SQUARE_TESTS) return;
@@ -1323,7 +1286,7 @@ public class RandomIndexerSampleTests {
       // This is the easy path through the method, even one sample
       // should be sufficient with n=3 and w=2.
       for (int i = 0; i < 3; i++) {
-        int[] result = RandomIndexer.nextWindowedIntTriple(3, 2, (int[]) null, true, gen);
+        int[] result = RandomIndexer.nextSortedWindowedIntTriple(3, 2, (int[]) null, gen);
         assertEquals(3, result.length);
         assertEquals(0, result[0]);
         assertEquals(1, result[1]);
@@ -1336,7 +1299,7 @@ public class RandomIndexerSampleTests {
       int n = 1000;
       int w = n - 1;
       for (int i = 0; i < 30; i++) {
-        int[] result = RandomIndexer.nextWindowedIntTriple(n, w, (int[]) null, true, gen);
+        int[] result = RandomIndexer.nextSortedWindowedIntTriple(n, w, (int[]) null, gen);
         assertEquals(3, result.length);
         assertTrue(result[0] < result[1]);
         assertTrue(result[1] < result[2]);
@@ -1344,7 +1307,7 @@ public class RandomIndexerSampleTests {
       n = 10000;
       w = n - 1;
       for (int i = 0; i < 30; i++) {
-        int[] result = RandomIndexer.nextWindowedIntTriple(n, w, (int[]) null, true, gen);
+        int[] result = RandomIndexer.nextSortedWindowedIntTriple(n, w, (int[]) null, gen);
         assertEquals(3, result.length);
         assertTrue(result[0] < result[1]);
         assertTrue(result[1] < result[2]);
@@ -1352,7 +1315,7 @@ public class RandomIndexerSampleTests {
       n = 100;
       w = n - 1;
       for (int i = 0; i < 30; i++) {
-        int[] result = RandomIndexer.nextWindowedIntTriple(n, w, (int[]) null, true, gen);
+        int[] result = RandomIndexer.nextSortedWindowedIntTriple(n, w, (int[]) null, gen);
         assertEquals(3, result.length);
         assertTrue(result[0] < result[1]);
         assertTrue(result[1] < result[2]);
@@ -1360,7 +1323,7 @@ public class RandomIndexerSampleTests {
       n = 10;
       w = n - 1;
       for (int i = 0; i < 30; i++) {
-        int[] result = RandomIndexer.nextWindowedIntTriple(n, w, (int[]) null, true, gen);
+        int[] result = RandomIndexer.nextSortedWindowedIntTriple(n, w, (int[]) null, gen);
         assertEquals(3, result.length);
         assertTrue(result[0] < result[1]);
         assertTrue(result[1] < result[2]);
@@ -1378,7 +1341,7 @@ public class RandomIndexerSampleTests {
         assertTrue(result[0] >= 0);
         assertTrue(result[2] < n);
         assertTrue(result[2] - result[0] <= w);
-        result = RandomIndexer.nextWindowedIntTriple(n, w, (int[]) null, true, gen);
+        result = RandomIndexer.nextSortedWindowedIntTriple(n, w, (int[]) null, gen);
         assertEquals(3, result.length);
         assertNotEquals(result[0], result[1]);
         assertNotEquals(result[0], result[2]);
@@ -1442,9 +1405,9 @@ public class RandomIndexerSampleTests {
     actual = RandomIndexer.nextWindowedIntTriple(5, 3, new int[2], gen);
     assertEquals(3, actual.length);
 
-    actual = RandomIndexer.nextWindowedIntTriple(5, 3, expected, false, gen);
+    actual = RandomIndexer.nextSortedWindowedIntTriple(5, 3, expected, gen);
     assertTrue(expected == actual);
-    actual = RandomIndexer.nextWindowedIntTriple(5, 3, new int[2], false, gen);
+    actual = RandomIndexer.nextSortedWindowedIntTriple(5, 3, new int[2], gen);
     assertEquals(3, actual.length);
   }
 
@@ -1464,7 +1427,7 @@ public class RandomIndexerSampleTests {
         assertTrue(min >= 0);
         assertTrue(max < n);
         assertTrue(max - min <= w);
-        result = RandomIndexer.nextWindowedIntTriple(n, w, true);
+        result = RandomIndexer.nextSortedWindowedIntTriple(n, w);
         assertNotEquals(result.i(), result.j());
         assertNotEquals(result.i(), result.k());
         assertNotEquals(result.k(), result.j());
@@ -1525,7 +1488,7 @@ public class RandomIndexerSampleTests {
       // This is the easy path through the method, even one sample
       // should be sufficient with n=3 and w=2.
       for (int i = 0; i < 3; i++) {
-        IndexTriple result = RandomIndexer.nextWindowedIntTriple(3, 2, true, gen);
+        IndexTriple result = RandomIndexer.nextSortedWindowedIntTriple(3, 2, gen);
         assertEquals(0, result.i());
         assertEquals(1, result.j());
         assertEquals(2, result.k());
@@ -1535,30 +1498,30 @@ public class RandomIndexerSampleTests {
       // go through the alternate path from above. That alternate path has
       // 3 potential subpaths, all approximately equally likely.
       int n = 1000;
-      int w = n - 1;
-      for (int i = 0; i < 30; i++) {
-        IndexTriple result = RandomIndexer.nextWindowedIntTriple(n, w, true, gen);
+      int w = n - 2;
+      for (int i = 0; i < 60; i++) {
+        IndexTriple result = RandomIndexer.nextSortedWindowedIntTriple(n, w, gen);
         assertTrue(result.i() < result.j());
         assertTrue(result.j() < result.k());
       }
       n = 10000;
-      w = n - 1;
-      for (int i = 0; i < 30; i++) {
-        IndexTriple result = RandomIndexer.nextWindowedIntTriple(n, w, true, gen);
+      w = n - 2;
+      for (int i = 0; i < 60; i++) {
+        IndexTriple result = RandomIndexer.nextSortedWindowedIntTriple(n, w, gen);
         assertTrue(result.i() < result.j());
         assertTrue(result.j() < result.k());
       }
       n = 100;
-      w = n - 1;
-      for (int i = 0; i < 30; i++) {
-        IndexTriple result = RandomIndexer.nextWindowedIntTriple(n, w, true, gen);
+      w = n - 2;
+      for (int i = 0; i < 60; i++) {
+        IndexTriple result = RandomIndexer.nextSortedWindowedIntTriple(n, w, gen);
         assertTrue(result.i() < result.j());
         assertTrue(result.j() < result.k());
       }
       n = 10;
-      w = n - 1;
-      for (int i = 0; i < 30; i++) {
-        IndexTriple result = RandomIndexer.nextWindowedIntTriple(n, w, true, gen);
+      w = n - 2;
+      for (int i = 0; i < 60; i++) {
+        IndexTriple result = RandomIndexer.nextSortedWindowedIntTriple(n, w, gen);
         assertTrue(result.i() < result.j());
         assertTrue(result.j() < result.k());
       }
@@ -1575,7 +1538,7 @@ public class RandomIndexerSampleTests {
         assertTrue(min >= 0);
         assertTrue(max < n);
         assertTrue(max - min <= w);
-        result = RandomIndexer.nextWindowedIntTriple(n, w, true, gen);
+        result = RandomIndexer.nextSortedWindowedIntTriple(n, w, gen);
         assertNotEquals(result.i(), result.j());
         assertNotEquals(result.i(), result.k());
         assertNotEquals(result.k(), result.j());